Move new classes into the proper locations
